+++ SpinIDE/TextHighlighting/SpinBackgroundHighlighter.m
@@ -72,7 +72,7 @@ typedef enum {tCON, tVAR, tOBJ, tPUB, tPRI, tDAT, tEOF, tNone} tokenType;
             // Color the block completed by the current token.
             if (token == tEOF)
                 // This handles files that don't end in a line feed, making sure the last line is highlighted.
-                lineStart = text.length;
+                lineStart = (int) text.length;
             NSRange textRange = {lastLineStart, lineStart - lastLineStart};
             
             ColoredRange *coloredRange = [[ColoredRange alloc] init];
